+++ versebot/versebot.py
@@ -26,6 +26,7 @@ try:
     kjvbible = pickle.load(open(configloader.getKJV(), 'rb'))
     nrsvbible = pickle.load(open(configloader.getNRSV(), 'rb'))
     drabible = pickle.load(open(configloader.getDRA(), 'rb'))
+    brentonbible = pickle.load(open(configloader.getBrenton(), 'rb'))
     print('Bible translations successfully loaded!')
 except:
     print('Error while loading Bible translations. Make sure the environment vars point to correct paths.')
@@ -78,7 +79,7 @@ while True:
                     lookupList.append(str(nextVer))
 
                 if len(lookupList) != 0:
-                    nextComment = constructComment(lookupList, comment, nivbible, esvbible, kjvbible, nrsvbible, drabible)
+                    nextComment = constructComment(lookupList, comment, nivbible, esvbible, kjvbible, nrsvbible, drabible, brentonbible)
                 else:
                     nextComment = False